refactor(gameEngine): Separate adaptive logic from game loop
- Remove methods like `isCryptoAvaliable`, `randomAttack`, `adaptiveAttack` and `counterMove` from `gameEngine`
- Add the same methods to the new `adaptiveEngine` as Separation of concerns
